% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display_NACol_order.R
\name{Display_NACol_order}
\alias{Display_NACol_order}
\title{Display the NA information in a data frame
Display NA percentages of columns in descending order}
\usage{
Display_NACol_order(input)
}
\arguments{
\item{input}{is a data frame}
}
\value{
a data frame with 2 columns: variable name, NA percentage
}
\description{
Display the NA information in a data frame
Display NA percentages of columns in descending order
}
