% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_tibble.R
\name{Impute_tibble}
\alias{Impute_tibble}
\title{Impute an all-character tibble}
\usage{
Impute_tibble(input)
}
\arguments{
\item{input}{is a tibble containing NA's}
}
\value{
is a NA-free tibble
}
\description{
Impute an all-character tibble
}
