% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_tibble_OneStop.R
\name{Impute_tibble_OneStop}
\alias{Impute_tibble_OneStop}
\title{A one stop imputation device for a tibble}
\usage{
Impute_tibble_OneStop(input)
}
\arguments{
\item{input}{a tibble with NA's}
}
\value{
a tibble with all NA's imputed
}
\description{
Impute numeric variables with "mice" function from "mice" packge
Impute character variables with corresponding modes
}
