% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_vector.R
\name{Impute_vector}
\alias{Impute_vector}
\title{Impute a vector with a designated value}
\usage{
Impute_vector(input1, input2)
}
\arguments{
\item{input1}{is the one-column data frame to be imputed}

\item{input2}{is the designated value to impute input1}
}
\value{
return a one-column data frame free of NAs
}
\description{
Impute a vector with a designated value
}
