% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_NumbyChr.R
\name{Replace_NumbyChr}
\alias{Replace_NumbyChr}
\title{Replaces all numbers in a character string by letters}
\usage{
Replace_NumbyChr(input)
}
\arguments{
\item{input}{is a character string with numbers}
}
\value{
a character without numbers
}
\description{
Replaces all numbers in a character string by letters
}
